Various small fixes (#770)
Fixes: 
- saleor auth client falling in loop when refresh token is expired
- "billing address not set" error being thrown within payment in some
cases
- summary closing unexpectedly after opening on mobile view
- remove some leftover console.logs

Also added a skeleton for delivery methods. Currently when logged in
user has no shipping address set but the auto selection is in progress,
the section says "please fill in the shipping address first" - while
that makes sense for the guest user, for signed in one it's already
happening so it makes more sense to replace the bit with skeleton
that'll actually indicate loading to user instead of requirement for
action.